+++ NEWS.md
@@ -1,5 +1,8 @@
 # matlib 0.9.0
 
+- `showEqn()` gains a `reduced` logical to print only the unique regression equations
+  when a `lm()`-type object is passed. May be combined with the `simplify` logical argument
+  for further reductions
 - `echelon()` gains a `reduced` logical to indicate whether the reduced or non-reduced form is computed
 - `powerMethod()` gains a `plot = TRUE` logical to draw the iteration history
 - added support for `'lm'` objects to `showEqn()` to show the design matrix equations